refactor(crucible): remove redundant runtime module prefixes
