+++ docs/audits/2026-08-24-remaining-roadmap-commonality-audit.md
@@ -76,7 +76,7 @@ candidates include dynamic registrations and conventional `main()` functions.
 | RR-008 | The five compatibility facade files repeat a four-launcher shape. | Exact census: package `__init__` plus `campaign`, `prepare`, `train`, and `watch_campaign` launchers. Artifact and installed-wheel checks pin this exact surface. | **Delete, do not abstract.** A shared facade factory would add code immediately before R8.4 removes the surface. Keep all five byte-stable through R8.3/R8.2. | R8.4, after `REL-004` and `STORE-003`. |
 | RR-009 | `scripts/slop_audit.py` omits the product package. | `SCAN_ROOTS` includes `core`, `plugins`, removed/absent `autoresearch`, and `scripts`, but not `geode_product`. | Replace the stale root with `geode_product` and pin the root census in the script's own focused test. The tool remains diagnostic, not a promotion oracle. | **Resolved:** exact production-root census is executable-tested. |
 | RR-010 | Two web-search tool classes duplicate the same dispatch/error/result body. | `WebSearchTool.aexecute` and `GeneralWebSearchTool.aexecute` differ materially only in tool name/description. Only `general_web_search` is in `definitions.json`; the former remains a `ToolRegistry` compatibility surface. | First decide whether the compatibility name is still public. Prefer deletion/convergence; only if both names must remain, extract one function rather than a base class. | **Resolved:** both public names remain and delegate to one function; behavior tests pin routing and result parity. |
-| RR-011 | Crucible repeats canonical JSON hashing and strict scalar validators. | At least seven modules repeat the same sorted compact JSON plus SHA-256 body; `contract.py` already has a canonical JSON helper, while `power.py` and `runtime_budget.py` repeat strict-field/text/number validators. | Reuse one existing Crucible contract helper after caller-specific error semantics are characterized. Do not introduce a generic repository-wide serialization framework. | **Resolved:** generic canonical hashes and exact validator copies share Crucible contract helpers; specialized normalization remains local. |
+| RR-011 | Crucible repeats canonical JSON hashing and strict scalar validators. | At least seven modules repeat the same sorted compact JSON plus SHA-256 body; `contract.py` already has a canonical JSON helper, while `power.py` and `admission/budget.py` repeat strict-field/text/number validators. | Reuse one existing Crucible contract helper after caller-specific error semantics are characterized. Do not introduce a generic repository-wide serialization framework. | **Resolved:** generic canonical hashes and exact validator copies share Crucible contract helpers; specialized normalization remains local. |
 | RR-012 | The self-improving CLI renders a product-visible placeholder. | `_render_preflight()` prints `harness no-op (PR-OPS-2b wires autoresearch/petri_raw)` while the runner deliberately uses `rerun_enabled=False`. | Replace the stale future-PR wording with the truthful current contract, or remove the line. Do not silently enable paid measurement from an interactive mutation command. | **Resolved:** preflight reports that measurement is skipped for interactive mutation. |
 
 ### 3.1 Non-roadmap cleanup evidence
