+++ evolve/crucible/admission/pilot.py
@@ -9,12 +9,12 @@ from pathlib import Path
 from typing import Any, Literal
 
 from evolve.crucible.artifacts import load_json_object
-from evolve.crucible.assays.runtime_receipt import load_runtime_receipt, runtime_artifact_bindings
+from evolve.crucible.assays.receipt import load_runtime_receipt, runtime_artifact_bindings
 from evolve.crucible.contract import ContractError, ExperimentContract
 from evolve.crucible.evidence import EvidenceEnvelope, expected_pairs, validate_evidence_identity
 
-from .runtime_budget import RUNTIME_ACCOUNTING_METHOD, RUNTIME_PILOT_SCHEMA
-from .runtime_identity import runtime_regime, runtime_regime_id
+from .budget import RUNTIME_ACCOUNTING_METHOD, RUNTIME_PILOT_SCHEMA
+from .regime import runtime_regime, runtime_regime_id
 
 
 def _raw_durations(path: Path, field: str) -> tuple[str, dict[tuple[str, int], float]]:
