+++ site/src/data/geode/changelog.ts
@@ -19,7 +19,7 @@ export const CHANGELOG: ChangelogEntry[] = [
   {
     "version": "Unreleased",
     "date": "",
-    "body": "### Architecture\n\n- **Evaluation and Crucible package trees now expose runtime ownership.**\n  MCPMark and tau2 integrations are vertical benchmark packages, Harbor lives\n  under external platforms, and Crucible separates admission, search, trusted\n  assays, and attestation while retaining v1.0.x adapter and persisted\n  provenance facades.\n- **Benchmark integrations and platform adapters now have distinct roles.**\n  MCPMark and tau2 remain pinned benchmark sources, while Harbor remains a\n  platform adapter outside the benchmark registry; research guidance now\n  requires Prime Agent and original/upstream parity checks when relevant.\n  Existing v1.0.x harness-named identifiers remain compatible but legacy.\n\n### Added\n\n- **Transparent cross-host skill inventory and security contracts.** All\n  repository development/meta skills now have one reviewed `.agents` source\n  with relative Claude Code aliases, bundled runtime skills are inventoried\n  separately, benchmark hashes are explicitly non-secret artifact identities,\n  and anti-deception secret checks no longer echo matched values into logs.\n- **Evaluation learning-view contract v2.** GEODE now validates digest-bound\n  `example → rollout → trajectory → reward` datasets while preserving native\n  receipts and the immutable `geode.trajectory@1` schema.\n- **Harbor terminal platform adapter.** The current GEODE AgenticLoop can run\n  as a Harbor external agent while preserving native verifier output and a\n  normalized GEODE trajectory sidecar.\n- **Native evaluation projector.** Harbor and tau2 result bundles can be\n  projected into the validated v2 learning view without rewriting native\n  receipts or collapsing infrastructure-invalid retries into reward rows."
+    "body": "### Architecture\n\n- **Crucible runtime modules now rely on package context instead of repeated\n  filename prefixes.** Admission uses `regime`, `budget`, `forecast`, and\n  `pilot`, while assay execution owns `receipt`; runtime artifact and schema\n  names remain unchanged.\n- **Evaluation and Crucible package trees now expose runtime ownership.**\n  MCPMark and tau2 integrations are vertical benchmark packages, Harbor lives\n  under external platforms, and Crucible separates admission, search, trusted\n  assays, and attestation while retaining v1.0.x adapter and persisted\n  provenance facades.\n- **Benchmark integrations and platform adapters now have distinct roles.**\n  MCPMark and tau2 remain pinned benchmark sources, while Harbor remains a\n  platform adapter outside the benchmark registry; research guidance now\n  requires Prime Agent and original/upstream parity checks when relevant.\n  Existing v1.0.x harness-named identifiers remain compatible but legacy.\n\n### Added\n\n- **Transparent cross-host skill inventory and security contracts.** All\n  repository development/meta skills now have one reviewed `.agents` source\n  with relative Claude Code aliases, bundled runtime skills are inventoried\n  separately, benchmark hashes are explicitly non-secret artifact identities,\n  and anti-deception secret checks no longer echo matched values into logs.\n- **Evaluation learning-view contract v2.** GEODE now validates digest-bound\n  `example → rollout → trajectory → reward` datasets while preserving native\n  receipts and the immutable `geode.trajectory@1` schema.\n- **Harbor terminal platform adapter.** The current GEODE AgenticLoop can run\n  as a Harbor external agent while preserving native verifier output and a\n  normalized GEODE trajectory sidecar.\n- **Native evaluation projector.** Harbor and tau2 result bundles can be\n  projected into the validated v2 learning view without rewriting native\n  receipts or collapsing infrastructure-invalid retries into reward rows."
   },
   {
     "version": "1.0.26",
