+++ tests/evolve/crucible/test_sealed.py
@@ -13,7 +13,7 @@ from typing import Any, Literal
 import evolve.crucible.attestation.sealed as sealed_module
 import pytest
 from evolve.crucible.artifacts import write_exclusive_json
-from evolve.crucible.assays.runtime_receipt import SharedRuntimeDeadline, runtime_artifact_bindings
+from evolve.crucible.assays.receipt import SharedRuntimeDeadline, runtime_artifact_bindings
 from evolve.crucible.attestation.bundle import PromotionBundle
 from evolve.crucible.attestation.sealed import (
     CorePromotionDecision,
